Initial RHEV network implementation
There is not much in this patch, it adds only network listing.
RHEV networking is quite different from cloud providers like EC2.
A user is allowed to create a VNIC(s) for a VM, then by associating
these VNICs with networks she specifies to which network the VM
belongs to.